\name{sbfc-package}
\alias{sbfc-package}
\alias{sbfc}
\docType{package}
\title{
\packageTitle{sbfc}
}
\description{
\packageDescription{sbfc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sbfc}
\packageIndices{sbfc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sbfc}

Maintainer: \packageMaintainer{sbfc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
