% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbfc.R
\name{signal_var_prop}
\alias{signal_var_prop}
\title{Signal variable proportion}
\usage{
signal_var_prop(sbfc_result, subset = F, thin = 50, nvars = 10,
  label_size = 1)
}
\description{
Computes proportion of the samples in which each variable is in the signal group (Group 1). Plots the first (nvars) variables in decreasing order of proportion.
}

